Prevent clicks and CPU usage improvement for VSTs
Changed to allocating a new voice for each note using round robin voice
stealing.
Added a silence detection method to free voices that are released and
quiet.
Fixed multiple notes occurring during a chunk so they each start at
their own delta.

+++ architecture/faust/vst/vstui.h
@@ -267,10 +267,9 @@ public:
 	/*
 		Return metadata for control (such as style, unit, etc...).
 	*/
-	const char* getControlMetadata(int index, const char* key, 
-																 const char* defaultString)
+	const char* getControlMetadata(int index, const char* key, const char* defaultString)
 	{
-		if (index < 0 || index > fUITable.size()) {
+		if (index < 0 || index > (int)fUITable.size()) {
 			TRACE(fprintf(stderr, "Illegal index (%d) accessed by getControlMetadata\n",
 										index));
 			return defaultString;
@@ -295,7 +294,7 @@ public:
 			return;
 		}
 
-		if (anyIndex >= fUITable.size()) {
+		if (anyIndex >= (int)fUITable.size()) {
 			TRACE(fprintf(stderr,"*** Faust vsti: %sIndex = %d too large!\n",
 						str,anyIndex));
 			return;
